+++ source-internet-banking/src/app/demo/pages/profile/profile.component.ts
@@ -83,8 +83,10 @@ export class ProfileComponent implements OnInit {
                     alert("Error. Please again!!");
                   });
               } else {
-                localStorage.clear();
-                this.router.navigateByUrl("/auth/signin");
+                if(confirm('Session has been expired. Please re-login.')){
+                  localStorage.clear();
+                  this.router.navigateByUrl("/auth/signin");
+                }
               }
             }
           );
